requirements.txt
# EcoPuntos - Dependencias del Proyecto
# Framework principal
Django==5.2.1

# Base de datos
psycopg2-binary==2.9.11  # Para PostgreSQL
dj-database-url==2.1.0  # Para configuración de DATABASE_URL en Railway

# Formularios y UI
django-crispy-forms==2.1
crispy-bootstrap5==2024.2

# Integración con Supabase (eliminada) - usar sólo PostgreSQL
# supabase==2.3.4

# Manejo de imágenes
Pillow==12.0.0

# Variables de entorno
python-decouple==3.8

# API REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Dependencias para setuptools/pkg_resources (necesario para Railway)
# Pin a <81 para mantener pkg_resources disponible (evita el aviso y rompe futuras versiones)
setuptools==79.0.1

# CORS para API
django-cors-headers==4.3.1

# Cache
django-redis==5.4.0
redis==5.0.1

# WebSockets
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0

# Logging y monitoreo
django-extensions==3.2.3

# Testing
pytest==7.4.4
pytest-django==4.8.0
pytest-cov==4.1.0

# Seguridad
django-ratelimit==4.1.0

# Producción
whitenoise==6.6.0
# gunicorn removed for Daphne-only production (use daphne as ASGI server)

# Desarrollo
django-debug-toolbar==4.2.0

# IA y Chatbot - Solo Gemini
google-generativeai==0.3.2  # Gemini 1.5 Flash - 1M tokens, gratis hasta 1.5K req/día
python-dotenv==1.0.1
# Almacenamiento de archivos en S3/Spaces
django-storages[boto3]==1.14.1
boto3==1.28.67